Miles + some cleaning (Fabric)
The name "KILOMETERS_PER_HOUR" goes off the setting page so I reduced the names down to "KILOMETER_PH" and "METERS_PS" which should suffice. Switched from ticks++ to MC.player.age for consistency with the forge version as that is the equivalent field. Added miles per hour as a mode. Also changed from initializing the string as "INVALID" to using a "default" in the switch.